requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
executing==2.2.0
fonttools==4.59.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-mcp-adapters==0.1.7
langchain-openai==0.3.25
langchain-text-splitters==0.3.8
langgraph==0.4.10
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.2.3
langgraph-sdk==0.1.70
langsmith==0.4.2
markdown-it-py==3.0.0
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mcp==1.10.0
mdurl==0.1.2
multidict==6.6.2
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.1
openai==1.91.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.3.6
stack-data==0.6.3
starlette==0.47.1
tenacity==9.1.2
tiktoken==0.9.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
